requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.4
httpx==0.28.1
aiofiles==24.1.0
pymongo==4.7.2
python-dotenv==1.1.0
tqdm==4.67.1
h2==4.2.0